add cloudinary functionality in all tabs

+++ src/app/api/review/[id]/route.ts
@@ -1,5 +1,12 @@
 import prisma from "@/utils/prisma";
 import { NextResponse } from "next/server";
+import { v2 as cloudinary } from "cloudinary";
+
+cloudinary.config({
+  cloud_name: process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
 
 export async function DELETE(
   request: Request,
@@ -7,12 +14,24 @@ export async function DELETE(
 ) {
   try {
     const { id } = params;
-    console.log(id);
+    const review = await prisma.reviews.findUnique({
+      where: { id },
+    });
+
+    if (!review)
+      return NextResponse.json(
+        { message: "Review not found" },
+        { status: 404 },
+      );
+
+    if (review.image) {
+      await cloudinary.api.delete_resources([review.image], {
+        type: "upload",
+        resource_type: "image",
+      });
+    }
     await prisma.reviews.delete({ where: { id } });
-    return NextResponse.json(
-      { message: "Reveiw deleted" },
-      { status: 200 },
-    );
+    return NextResponse.json({ message: "Reveiw deleted" }, { status: 200 });
   } catch (error) {
     console.error("Error in DELETE /review:", error);
     return NextResponse.json(
